glass: move local node status call to proper service
